:recycle: refactor: 북마크 service 수정
